Finding a Public Agent IP - Improve Oneliner
Add `and .attributes.public_ip == "true"` to the `select(...)` so that we get just the public agents, not the public agents and masters.

I'm new to dc/os so I'm not actually sure this is guaranteed to work forever, as I'm not that familiar with the promises of the `dcos node` api, but it seems to be working for me. Comments are appreciated. Maybe the correct fix is actually to make sure `slave_public` is null for masters? That, however, would be a code change for which I'm certainly not qualified.